Refactor: add modal template story
- correct: remove unused code
