+++ src/components/Modals/CardDetail/style.ts
